переименовал script.js в popup.js
