+++ app/catalog.html
@@ -434,6 +434,6 @@
 		<script type="module" src="js/getUrlParam.js"></script>
 		<script type="module" src="js/catalog.js"></script>
 
-		<script type="module" src="js/script.js"></script>
+		<script type="module" src="js/popup.js"></script>
 	</body>
 </html>
